sat/requirements.txt
deepspeed>=0.15.1
transformers>=4.45.0
xformers>=0.0.28
torch>=2.4.0
pytorch_lightning>=2.4.0
torchvision>=0.19.0
einops>=0.8.0
fsspec>=2024.6.1
kornia>=0.7.3
numpy>=2.1.1
omegaconf>=2.3.0
open_clip_torch>=2.26.1
Pillow>=10.4.0
safetensors>=0.4.5
scipy>=1.14.1
SwissArmyTransformer>=0.4.12
tqdm>=4.66.5
wandb>=0.18.1
openai>=1.48.0